Fix undefined behavior and memory safety bugs in SimulationRuntime/c
- Add missing #include <string.h> to stateset.c
- Fix fclose(NULL) UB in MoveData.c and simulation_input_xml.c
- Fix overlapping strcpy -> memmove in utility.c:362
- Fix sizeof(modelica_real*) -> sizeof(modelica_real) in newton_diagnostics.c
- Fix memory leak: free x0, f, dx, fx, fxx, n_idx on early return in
  newton_diagnostics.c when p == 0
- Fix integer division (i/n always 0) -> (double)i/n in
  nonlinearSolverHomotopy.c initial guess perturbation
- Fix fscanf "%s" buffer overflow -> "%199s" in MoveData.c
- Fix NULL deref after fopen: add NULL checks in DebugeOptimization.c
  (4 locations) and nonlinearSystem.c

+++ OMCompiler/SimulationRuntime/c/simulation/solver/nonlinearSolverHomotopy.c
@@ -2295,13 +2295,13 @@ NLS_SOLVER_STATUS solveHomotopy(DATA *data, threadData_t *threadData, NONLINEAR_
       {
         debugString(OMC_LOG_NLS_V, "assert handling:\t vary initial guess by +1%.");
         for(i = 0; i < homotopyData->n; i++)
-          homotopyData->x0[i] = homotopyData->xStart[i] + homotopyData->xScaling[i]*i/homotopyData->n*0.01;
+          homotopyData->x0[i] = homotopyData->xStart[i] + homotopyData->xScaling[i]*(double)i/homotopyData->n*0.01;
       }
       if (tries == 2)
       {
         debugString(OMC_LOG_NLS_V,"assert handling:\t vary initial guess by +10%.");
         for(i = 0; i < homotopyData->n; i++)
-          homotopyData->x0[i] = homotopyData->xStart[i] + homotopyData->xScaling[i]*i/homotopyData->n*0.1;
+          homotopyData->x0[i] = homotopyData->xStart[i] + homotopyData->xScaling[i]*(double)i/homotopyData->n*0.1;
       }
     }
     if (success != NLS_SOLVED) {
